packages/molmcp-molpack/pyproject.toml
[project]
name = "molmcp-molpack"
version = "0.1.0"
description = "MolCrafts MCP plugin: molpack packing-script inspector"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}

authors = [
    { name = "Roy Kid", email = "lijichen365@gmail.com" }
]

keywords = ["mcp", "model-context-protocol", "molcrafts", "molpack", "packmol"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "fastmcp>=3.0",
    "molcrafts-molmcp",
    "molcrafts-molpack",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
molmcp-molpack = "molmcp_molpack.cli:main"

[tool.uv.sources]
molcrafts-molmcp = { workspace = true }
molcrafts-molpack = { path = "../../../molpack/python", editable = true }

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
